eye_disease_detection/requirements.txt
Django==4.2.7
tensorflow==2.19.0
Pillow==10.0.1
numpy
opencv-python==4.8.1.78